requirements.txt
flask
pandas
openpyxl
qrcode
pillow
gunicorn
gspread
oauth2client
